BUG 35895313 - [35889353->23.09.1] Topic subscriber channel allocations are not always correct after member departure (merge ce/main -> ce/23.09 103849)
[git-p4: depot-paths = "//dev/coherence-ce/release/coherence-ce-v23.09/": change = 103850]

+++ prj/coherence-docker/push-images.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# Copyright (c) 2000, 2022, Oracle and/or its affiliates.
+# Copyright (c) 2000, 2023, Oracle and/or its affiliates.
 #
 # Licensed under the Universal Permissive License v 1.0 as shown at
 # https://oss.oracle.com/licenses/upl.
@@ -18,25 +18,16 @@ if [ "${NO_DAEMON}" != "true" ]
 then
   buildah pull "docker-daemon:${IMAGE_NAME}-amd64"
   buildah pull "docker-daemon:${IMAGE_NAME}-arm64"
-  buildah pull "docker-daemon:${IMAGE_NAME}-graal-amd64"
-  buildah pull "docker-daemon:${IMAGE_NAME}-graal-arm64"
+  if [ "${NO_GRAAL}" != "true" ]
+  then
+    buildah pull "docker-daemon:${IMAGE_NAME}-graal-amd64"
+    buildah pull "docker-daemon:${IMAGE_NAME}-graal-arm64"
+  fi
 fi
-
-# If the registry is docker.io then cut the registry from the front of the image name
-# e.g. docker.io/foo/bar:1.0 becomes foo/bar:1.0
-# This is because Buildah's local name of an image cuts docker.io from the name
-REGISTRY=$(echo "${IMAGE_NAME}" | cut -d"/" -f1)
-if [ "${REGISTRY}" == "docker.io" ]
-then
-  IMAGE_SUFFIX=$(echo "${IMAGE_NAME}" | cut -d"/" -f2-10)
-else
-  IMAGE_SUFFIX="${IMAGE_NAME}"
-fi
-LOCAL_NAME="localhost/${IMAGE_SUFFIX}"
 
 buildah images
 
-# Create distroless based images
+buildah rmi "${IMAGE_NAME}" || true
 buildah manifest rm "${IMAGE_NAME}" || true
 buildah manifest create "${IMAGE_NAME}"
 buildah manifest add --arch amd64 --os linux "${IMAGE_NAME}" "${IMAGE_NAME}-amd64"
@@ -45,14 +36,17 @@ buildah manifest inspect "${IMAGE_NAME}"
 
 buildah manifest push --all -f v2s2 "${IMAGE_NAME}" "docker://${IMAGE_NAME}"
 
-# Create Graal based images
-buildah manifest rm "${IMAGE_NAME}-graal" || true
-buildah manifest create "${IMAGE_NAME}-graal"
-buildah manifest add --arch amd64 --os linux "${IMAGE_NAME}-graal" "${IMAGE_NAME}-graal-amd64"
-buildah manifest add --arch arm64 --os linux "${IMAGE_NAME}-graal" "${IMAGE_NAME}-graal-arm64"
-buildah manifest inspect "${IMAGE_NAME}-graal"
+if [ "${NO_GRAAL}" != "true" ]
+then
+  # Create Graal based images
+  buildah manifest rm "${IMAGE_NAME}-graal" || true
+  buildah manifest create "${IMAGE_NAME}-graal"
+  buildah manifest add --arch amd64 --os linux "${IMAGE_NAME}-graal" "${IMAGE_NAME}-graal-amd64"
+  buildah manifest add --arch arm64 --os linux "${IMAGE_NAME}-graal" "${IMAGE_NAME}-graal-arm64"
+  buildah manifest inspect "${IMAGE_NAME}-graal"
 
-buildah manifest push --all -f v2s2 "${IMAGE_NAME}-graal" "docker://${IMAGE_NAME}-graal"
+  buildah manifest push --all -f v2s2 "${IMAGE_NAME}-graal" "docker://${IMAGE_NAME}-graal"
+fi
 
 # prune all old images that are no longer associated to a tag
 buildah rmi --prune || true
